refactor: store 디렉토리 파일구조 변경
store 하위에 products.js carts.js 도메인별로 파일이 있고
각 파일에 해당 redux 관련 함수들을 모아둠(action, reducer 등)
store 를 index.jsx 에 선언해서 Provider로 제공
