+++ src/components/Main/ProductListContainer.jsx
@@ -2,7 +2,8 @@ import styled from 'styled-components';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import Product from 'components/Main/Product';
-import { loadCarts, loadProducts } from 'store/actions';
+import { loadCarts } from 'store/carts';
+import { loadProducts } from 'store/products';
 
 function ProductListContainer() {
   const { products } = useSelector((state) => state.products);
