+++ src/store/carts.js
@@ -1,18 +1,20 @@
 import axios from 'axios';
 import { SERVER_URL, PATH } from 'constants';
-import TYPE from 'store/types';
+
+const TYPE = {
+  CARTS_LOAD: 'carts/load',
+  CARTS_ADD: 'carts/add',
+  CARTS_DELETE: 'carts/delete',
+};
+
+const initialState = {
+  carts: [],
+};
 
 const actionCreators = {
-  loadProducts: (payload) => ({ type: TYPE.PRODUCTS_LOAD, payload }),
   loadCarts: (payload) => ({ type: TYPE.CARTS_LOAD, payload }),
   addCart: (payload) => ({ type: TYPE.CARTS_ADD, payload }),
   deleteCart: (payload) => ({ type: TYPE.CARTS_DELETE, payload }),
-};
-
-export const loadProducts = () => async (dispatch) => {
-  const { data } = await axios.get(`${SERVER_URL}${PATH.PRODUCTS}`);
-
-  dispatch(actionCreators.loadProducts(data));
 };
 
 export const loadCarts = () => async (dispatch) => {
@@ -40,5 +42,21 @@ export const deleteCart = (id) => async (dispatch) => {
     dispatch(actionCreators.deleteCart(id));
   } catch (error) {
     console.log('errorL ', error.messages);
+  }
+};
+
+export const cartsReducer = (state = initialState.carts, action) => {
+  switch (action.type) {
+    case TYPE.CARTS_LOAD:
+      return { ...state, carts: action.payload };
+    case TYPE.CARTS_ADD:
+      return { ...state, carts: state.carts.concat(action.payload) };
+    case TYPE.CARTS_DELETE:
+      return {
+        ...state,
+        carts: state.carts.filter((cart) => cart.id !== action.payload),
+      };
+    default:
+      return state;
   }
 };
