--- src/store/store.js
@@ -1,7 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import rootReducer from 'store/reducer';
-
-const store = createStore(rootReducer, applyMiddleware(thunk));
-
-export default store;
