--- src/store/types.js
@@ -1,8 +1,0 @@
-const TYPE = {
-  PRODUCTS_LOAD: 'products/load',
-  CARTS_LOAD: 'carts/load',
-  CARTS_ADD: 'carts/add',
-  CARTS_DELETE: 'carts/delete',
-};
-
-export default TYPE;
